requirements.txt
streamlit
chromadb
sentence-transformers
geopandas
shapely
rasterio
matplotlib
openai
pillow
folium
scikit-learn
requests